requirements.txt
wheel
click
netifaces
setuptools
dataclasses
cryptography
pyOpenSSL
